Rearrange crate structure
Now a proper rust library to be used to build a cdylib.
Disable some compiler naming warnings on linked names.
Includes an R script for manually testing generated bindings.
