+++ src/internals/c.rs
@@ -1,13 +1,16 @@
 use crate::internals::sexpr::{SexpType, Sexpr};
 
 // https://github.com/wch/r-source/blob/abb550c99b3927e5fc03d12f1a8e7593fddc04d2/src/main/arithmetic.c#L337
+#[allow(non_upper_case_globals)]
 pub const R_NaInt: i32 = std::i32::MIN;
 //not guarenteed, but commited 22+ years ago and not changed since.
 //https://github.com/wch/r-source/blame/abb550c99b3927e5fc03d12f1a8e7593fddc04d2/src/main/arithmetic.c#L115
+#[allow(non_snake_case)]
 pub fn R_NaReal() -> f64 {
     f64::from_bits(0x7ff00000u64 | 0x1954)
 }
 //https://github.com/wch/r-source/blob/5a156a0865362bb8381dcd69ac335f5174a4f60c/src/include/R_ext/Arith.h#L55
+#[allow(non_upper_case_globals)]
 pub const R_NaLogical: i32 = R_NaInt;
 #[link(name = "R")]
 extern "C" {
